requirements.txt
django==4.1.*
djangorestframework==3.14.*
django-environ==0.9.*
django-rest-knox==4.2.*
django-dynamic-preferences==1.14.*
mysqlclient==2.1.*
django-filter==22.1
drf-nested-routers==0.93.*